move hard-coded network setting in a .ini file, ignore media files
The network settings are now loaded from a 'network.ini' file where
different networks can be added by the user and instantiated with the
'net' argument. The config file is loaded from the script directory and
the paths within it are relative to the config file directory.
Ignore in GIT the media files supported by the demo.